add interface for metrics with state and updated r2
